discretize distance

+++ Crawler/Model/Action.cs
@@ -49,5 +49,5 @@ public abstract class Action
     }
 
     /// For AI and UI use. Range is inclusive.
-    public virtual (float min, float max) Range => (0, 0);
+    public virtual (int min, int max) Range => (0, 0);
 }
