+++ Crawler/Model/Attacks/ReachAttackData.cs
@@ -6,7 +6,7 @@ public abstract class ReachAttackData : Resource
     [Export] public int recovery = 5;
     [Export] public int stun = 10;
     [Export] public int damage = 2;
-    [Export] public float range = 2;
+    [Export] public int range = 2;
 
     public void DoAttack(Entity target, int timeNow)
     {
