+++ Helper/PathFinder.cs
@@ -31,7 +31,7 @@ public class PathFinder
     /// Makes the heuristic easy.
     public PathResult Run()
     {
-        Dictionary<(int, int), float> cost = new Dictionary<(int, int), float>();
+        Dictionary<(int, int), int> cost = new Dictionary<(int, int), int>();
         SimplePriorityQueue<(int, int)> frontier = new SimplePriorityQueue<(int, int)>();
         
         result.nextStepFor = new Dictionary<(int, int), (int, int)>();
@@ -68,7 +68,7 @@ public class PathFinder
                 // Filter neighbors. This search goes backwards, remember.
                 if (!walkable((neighbor, current))) { continue; }
 
-                float distance = GridHelper.Distance(current, neighbor);
+                int distance = GridHelper.Distance(current, neighbor);
                 if (!cost.ContainsKey(neighbor) || cost[current] + distance < cost[neighbor])
                 {
                     cost[neighbor] = cost[current] + distance;
